Generate resolver function types
Summary:
This diffs adds type generation for the resolver functions.

Next to the imported type of the resolver function we will generate an additional type assertion:

```
import myResolverFn from "myResolverFn"
(myResolverFn: (
  rootKey: myResolverFnFragmentKey$key,
  // we also generating flow-types for `args`.
) => mixed);
```

Now, if the resolver function doesn't match the generated type assertion there will be a flow error in the generated artifact.

- The bad thing about this: flow error in generated artifact may be confusing and not very clear on how to fix.
- The good thing about this: it will be impossible to flowfixme these types in generated artifacts.

Typescript is not implemented yet. But I suspect we can do something like this for TS:
https://www.typescriptlang.org/play?#code/GYVwdgxgLglg9mABADwBQEMBOBzAXIgZykxjGwEp8wQBbAIwFNNEBvAKEU8UwahEyQAGANxsAvmzYAbXogD6 (https://github.com/facebook/relay/commit/4173a5e149a073d792dbed48d32ede2365ba7b3e)+DDiq1GmcogC8APkTowATy0pRQA

Reviewed By: captbaritone

Differential Revision: D36677395

fbshipit-source-id: 7ec7ca3a9bc888341766a076a7d0333c0ca39d6d