Subscribe based on latest (missed) updates
Summary:
Updates the new React Cache version of `useFragmentInternal()` to handle edge cases in checking for missed updates. On commit (in useEffect), useFragment has to check for modifications to the store that occurred between when the hook read data and the time of commit and trigger a re-render. The previous version checked for changes to fragment *data* — triggering a re-render — but subscribed to the store using the stale snapshot created during render.

The edge case is that some changes to the store can affect *which* data should be subscribed to without changing the actual result of the fragment. In the new test case, the missed update changes a user's profile_picture object to point to a different record, but doesn't update the field (`uri`) that the fragment reads. Then, a subsequent update to the store updates the `uri` on the new profile_picture record. Previously useFragmentInternal would ignore this case since it subscribed to the old record; now it correctly updates.

Note that FragmentResource correctly handled this case, so this brings useFragmentInternal() up to parity.

*For a follow-up*, I realized while implementing this that are possibly other changes that can occur that may not be captured as changes to the result data. For example, it might be possible for a missing client edge to change as a result of a relationship change, without affecting the data (i'm not sure, have to double check). I added a todo for this since it needs more investigation and since FragmentResource also doesn't handle these cases.

Reviewed By: alunyov

Differential Revision: D36882905

fbshipit-source-id: 14e2c5a9bb702339bf198255d7a86ad8ad3d83fd