theoretical edge case for batched state changes
Summary:
Fixes a *theoretical* issue that can occur when different types of state updates are batched. When enqueuing a setState due to a change in data, useFragmentInternal now checks to see if the current state's selector has changed. If so, it bails out of the state update.

It's *really* hard to construct a unit tests for this case since it requires triggering a relay data change at precisely the right time relative to React's operation. However it seems safer to add this logic than not to, so i'm inclined to ship this w/o a test. Open to feedback here.

Reviewed By: davidmccabe

Differential Revision: D36988439

fbshipit-source-id: 648870657599d205c1de72f5c021a9078a6400c5