Correctly remove RefetchableFragment artifacts for ClientEdges in incremental mode.
Summary:
The root cause of the issue was an incorrect artifact mapping from the source to the generated artifacts for a fragment.

For `RefetchableFragments` generated for client-edge queries the source was the name of the virtually created fragment, where we should be using the original "client-edge" fragment spread source.

Interestingly, we already have a special case for these client-edge generated fragments, where their source hash (see the directive CLIENT_EDGE_GENERATED_FRAGMENT_KEY) was generated for the "client-edge".

This diff also passes this `source_name` to the reader's `source_definition_names`, which later will be used in the compiler's artifact map to correctly attribute generated artifacts to their source fragments.

Reviewed By: josephsavona

Differential Revision: D37116613

fbshipit-source-id: 3bccc96a5358dd53ab8f114407e07239f37a56c5