useClientQuery hook
Summary:
Based on the discussion here: https://github.com/facebook/relay/issues/3760#issuecomment-1156285034

This diff adds a new hook, that is meant to be used to render client-only queries. We're doing this to be intentional about our APIs. We want to make sure that client-only queries cannot be used with the API that can send a network request. In general, `FetchFn` with client-only query will throw, if request does not have a persisted query ID or a query text.

Implementation:
This hook is a just `useLazyLoadQuery` hook with `store-only` fetch policy.

The flow should enforce that client-only queries are used with the correct hook (ClientQuery type is removed for `useLazyLoadQuery`).

Reviewed By: josephsavona

Differential Revision: D37171439

fbshipit-source-id: 970664a025b361805aaff8fc39270d8c2f4ba837