Import type for resolver function in client edges
Summary:
Our typegen for client-edges was not generating import types for resolver fields. This was happening because we were removing the resolver filed for the client-edges in the typegen - to generate the correct selection type for the client-edge reader.

But we need both - client-edge selection for the reader, and also the resolver field - to import resolver function type assertion.

This diff:
- in the apply_transforms for typegen we will keep the client-edge fields unchanged.
- in the typegen visitor we need a special handling client-edges. For inline fragments with client-edge directive we will generate import type for their resolver function. And we will traverse into their fragment selections to keep reader types correct.

Reviewed By: captbaritone

Differential Revision: D37211218

fbshipit-source-id: 729fec25d85bc007e7c35c47ab28bf25cffd04cd