Add newtypes around Variable names V2
Summary:
Modify VariableDefinition's and Variable's name field to contain VariableName new type instead of a StringKey
In build.rs, (and wherever else these types are constructed, e.g. in various transforms)

This Diff is in continuation of https://www.internalfb.com/diff/D38388633
This also includes the changes in graphql_build_infra to support the change.

Reviewed By: rbalicki2

Differential Revision: D38556809

fbshipit-source-id: 65388bf2380f291c0b8eab7bac97c11210a04c3f