Default FetchPolicy for Live Queries to "store-and-network"
Summary:
If you have already fetched or prefetched a live query and you come back to the same page and it serves from cache, it doesn't "work" any more - the query is no longer "live". The solution is always to manually set the fetchPolicy to "store-and-network" or "network-only" - we should make this the default.

Technically maybe we should also *warn* if someone explicitly passes "store-or-network" here with a live query as it could cause bugs? But at least let's make the default behavior the correct one - for live queries we need to still do the network request to set up polling/live updates.

Reviewed By: kassens

Differential Revision: D30669268

fbshipit-source-id: 089d96343a81d9aa102ea21ef77e9cbc4b94e070