Make `derived_from` optional on SplitOperationMetadata
Summary:
This will allow us to generate split operations that are not derived from fragments. It will be used for $normalization operations generated for Relay resolvers that want to return objects.

In this specific diff we do not change the observable behavior of the compiler. But the `panic!` is moved from `split_operation_metadata` to `generate_artifacts`.

Reviewed By: rbalicki2

Differential Revision: D39144124

fbshipit-source-id: b6105fd1bd63e6b7ef62c5228fa1269c8b719f2b