Explicitly `jest.useFakeTimers('legacy')` when legacy timers are required
Summary:
In Jest 27+, calling `jest.useFakeTimers()` with no argument is equivalent to `jest.useFakeTimers('modern')`, which is a breaking change from Jest 26 where "legacy" timers are the default.

The difference in implementation and available fake timer APIs causes several of our tests to fail.

This diff updates the tests that would otherwise fail to use `useFakeTimers('legacy')` explicitly. This is a no-op change in our current Jest 26, but allows an update to Jest 27 without further modification of these tests.

Reviewed By: jacdebug

Differential Revision: D39258784

fbshipit-source-id: 54833ff640bb19be2ac3f38367d085e0a353b2a3