Strip extensions on file import paths when writing out TypeScript generated code (#4057)
Summary:
# What does this PR do

* This PR, imported from github, scrubs `.ts` extensions from filepaths in imports in generated typescript files
* Typescript does not like `.ts`: https://github.com/microsoft/TypeScript/issues/37582

In the typegen for Typescript when we're writing the pathname for module and type imports, this PR strips any trailing `.ts` that might exist on the path. This is required for TypeScript file imports which don't like the trailing file suffix. This was partially fixed in https://github.com/facebook/relay/pull/4002 but not completely.

Pull Request resolved: https://github.com/facebook/relay/pull/4057

Test Plan: :eyes:

Reviewed By: alunyov

Differential Revision: D39437828

Pulled By: rbalicki2

fbshipit-source-id: 53b3f5f7e467cefbc9ae73bf455ca040ea8a69aa