New docblock annotation @outputType
Summary:
This will mark resolver as the resolver that will return a complex object.

These will work similar to `edgeTo`, but instead of the `DataID` these resolves can return complex structures that matches the shape of the schema.

The schema extension, generated for `outputType` marked with directive relay_resolver with `is_output_type: true` argument. This argument will be used to generate $normalization artifacts for it.

In the following diffs we will add a transform that with this (normalization artifacts) code generation.

Reviewed By: rbalicki2

Differential Revision: D39111651

fbshipit-source-id: 8f41d4ff59c90226a936dce8f6125c7d52c7007e