Compiler flag to enable named imports
Summary:
If this feature flag is enabled, we'll use named imports for resolver functions. This will allow us define multiple resolvers in a single file.

In this diff we passing this flag to the relay docblock parser, so it can start adding field name (for now) as a `named` import.

In the next diff this `import_name` will be used in codegen/typegen.

Reviewed By: captbaritone

Differential Revision: D40240874

fbshipit-source-id: 7944ee65a52bcbcb6b5edc305c15ae6c2cce561f