Generate fragment for Relay Resolver Models
Summary:
Update initial version of `__self` fragment transform generation.

For types that have `__relay_model_instance` (this is a "reserved" name for Relay resolvers, and should not be selectable by users) we will generate fragments selecting this field.

Later,  the value of this `__relay_model_instance` will be passed to the individual resolvers defined on the Model (a version of this: D38260086)

Reviewed By: captbaritone

Differential Revision: D40276487

fbshipit-source-id: 458fb91cc6d573ad3bfd957c689d8025c35be3f2