Do not use path in snapshots, they are different on windows. (#4101)
Summary:
The main issue is that is causing the path to the resolver on the windows to be different is this:

https://github.com/Manishearth/pathdiff/blob/a78acde0955a556b2bdf11e6a9efae459c4bcf80/src/lib.rs#L75

In the compiler we try to construct a path from the generated artifact to the JS module, using this `pathdiff` utility. And it looks like on windows, when we construct this relative path the components are joined using an os-specific separator.

I believe this is correct: constructing an os-specific path. But it breaks the os-specific snapshot tests (and I don't think we have a good way of disabling these generated tests based on the OS, atm).

Additional concern for `pathdiff`, it may also create a problem when people may keep checked-in artifacts in the repo, and also use mixed dev environments (windows/Linux laptops) to run relay.

I think we may somehow address this issue in a separate PR, but this PR is just to unblock Rust build on CI, and fix: https://github.com/facebook/relay/issues/4089

Pull Request resolved: https://github.com/facebook/relay/pull/4101

Reviewed By: lumapat

Differential Revision: D40559826

Pulled By: alunyov

fbshipit-source-id: cca0d241998268bd3c2c5059029bf77d0765f7e3