Extract model resolvers (and extend schema with types) before handling other resolvers
Summary:
For Relay Model Resolvers we are generating object/scalar type schema extensions.

In order to use these types in field resolvers they need to be available on the schema, otherwise there will be compiler error, if you try to define resolver that returns the Model's type.

In this diff we're splitting the schema building for docblock into two phases. First, we extend schema with model resolvers.

Second, we'll extend the exiting types with "regular" resolver's extensions.

Reviewed By: captbaritone

Differential Revision: D40433600

fbshipit-source-id: b68497ad993ab92bae274db5f250615be61e0fd8