Resolvers that return @weak objects
Summary:
For resolvers that want to return weak objects we expect the output of the function to be
```
function weakObjectResolver<T>(...): ...
  return { __relay_resolver_instance: T }
}
```

But, this will leak implementation details to the product code:

- the fact that we're using outputType to make these weak objects resolvers
- and that we're using `__relay_resolver_instance` field to store the instance of this object.

As a developer you want to write your resolver as this:

```
function weakObjectResolve<T>(...): T
  return (something as T);
}
```

To support this API, and still rely on the behavior of the outptutType, we need to automate the wrapping of `T` into the object with instance key. The type generation also has to reflect that.

This diff is adding support for this:
- Make weak object types as `@__RelayWeakObject`
- In the codegen, pass resolver function to the HOF that will do the wrapping of the `T`
- In the typegen, update the generated types to expect `T` from the resolver, and not the `{__relay_resolver_instance: T}`

Reviewed By: captbaritone

Differential Revision: D40454293

fbshipit-source-id: 6527bed182d3650dc5e399910ed1db5023d7ae61