Remove optional initialQueryRef arg from use-query-loader.md example (#4054)
Summary:
As per [this stack overflow question](https://stackoverflow.com/questions/68708300/how-to-obtain-the-initialqueryref-in-relay), the docs are confusing on first glance. It looks like the user _needs_ to provide an `initialQueryRef` argument and the docs reference a router, but thats quite a bit out of context, and the argument isn't needed.

I imagine quite a few devs have looked at this example, couldn't figure out how to obtain the initialQueryRef and have given up.

This removes the reference from the code example for clarity sake and a follow-up PR should add a fully working example where the initial query ref is passed.

Pull Request resolved: https://github.com/facebook/relay/pull/4054

Reviewed By: alunyov

Differential Revision: D40707986

Pulled By: rbalicki2

fbshipit-source-id: cae7d7f40b63829f29c5a8bc03c4def72737a50e