Validate compiler output in OSS (#4116)
Summary:
This change will check the `git status` after we run the compiler on test project and build unit-tests.

The check will fail if there are changed files. This will ensure that the changes to config.test.json internally are in sync with the OSS config.test.json.

Pull Request resolved: https://github.com/facebook/relay/pull/4116

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

The CI actually detected outdated artifacts in the `test-project`:

https://github.com/alunyov/relay/actions/runs/3372454219/jobs/5595899864

Reviewed By: lumapat

Differential Revision: D40913806

Pulled By: alunyov

fbshipit-source-id: 1db03c5a96b5492fbf5f20b5586d8042132689e0