Add a new package eslint-plugin-relay-internal (#4147)
Summary:
Adds a new package that we can use to add lint rules that apply just to the Relay code base. It starts with `no-mixed-import-and-require` which should catch when we mix require and import, which is disallowed in WWW's Flow config and so breaks our sync.

Pushing now to see what problems crop up when imported into Phabricator. Will followup there with whatever work is needed to get this to work with internal linting.

Pull Request resolved: https://github.com/facebook/relay/pull/4147

Test Plan:
{F819037023}

Validate that D41857287 produces a Phabricator Signal error

Differential Revision: D41849549

Pulled By: captbaritone

fbshipit-source-id: fd500bdd6871c2713c1fca020eab5319e873a235