Add a new package eslint-plugin-relay-internal (#4147)
Summary:
Adds a new package that we can use to add lint rules that apply just to the Relay code base. It starts with `no-mixed-import-and-require` which should catch when we mix require and import, which is disallowed in WWW's Flow config and so breaks our sync.

Pushing now to see what problems crop up when imported into Phabricator. Will followup there with whatever work is needed to get this to work with internal linting.

Pull Request resolved: https://github.com/facebook/relay/pull/4147

Test Plan:
## In Open Source

I've confirmed that if I break the lint rule's tests they are caught by our top-level (Yarn workspace) jest run.

{F819060837}

 ---
I've also confirmed that the lint rule triggers both in the editor and at the CLI.

{F819060940}

{F819061082}

## In Xplat

I've confirmed that the lint rule triggers in the editor

{F819061495}

And on Phabricator (see D41857287)

{F819061807}

Reviewed By: davidmccabe

Differential Revision: D41849549

Pulled By: captbaritone

fbshipit-source-id: 3c0d6969705c6a506eeba62f8306515754015cee