Ensure imports always have a name
Summary:
Our way of modeling imports in the codegen AST allowed us to create imports that did not have a valid name to import as. In commonjs this worked because you can require something directly and assign the result directly without ever assigning it to a variable (in some cases). However to support ES Modules, we have to perform the import at the top of the file and assign it a variable name.

Luckily we already had a struct that expresses only valid dependencies `JSModuleDependency`. This diff pulls that structure into the AST itself and ensures every import has a name.

Reviewed By: alunyov

Differential Revision: D42795056

fbshipit-source-id: 68aa577ec95d04b5922caf6c048d1ae58d219e1f