Add isOutputType flag to resolver normalization node
Summary:
To correctly traverse into client edge selections we will need to know how to read the id of the related object out of the resolver record. This looks different for output type (which returns a fat object) than it does for edge to (which returns an id). Currently we handle this by looking for the presense of a specific field in the resolver record. This seems a bit error prone. This diff stack adds an additional normalization artifact field to tell us which type we expect (output type or not output type).

For now, we make the field optional from a type perspective so that we can land changes in this order:

1. Make isOutputType an optional field from a type perspective
2. Update the compiler to always include isOutputType
3. Make isOutputtype non-optional from a type perspective.
4. Update the runtime to depend upon having isOutputType

Reviewed By: rbalicki2

Differential Revision: D42690435

fbshipit-source-id: 17ca976ed35e6efb6857376588a8255ade1ccb5c