Only reference mark outputType resolvers when client edge
Summary: Previously every resolver would check itself to see if it was a client edge to an outputType object. Now that we have a client edge to client object normalization node we can move that logic to the edge node and only check in the case that we are actually an outputtype edge. This also sets the stage for handling edgeTo edges to strong objects in a later diff.

Reviewed By: rbalicki2

Differential Revision: D42690433

fbshipit-source-id: 9ca1c96d70a9ff8b925695da777547a531b897ff