Ensure graphql artifact imports all have names
Summary: Currently a dependency on a GraphQL module just specifies a string that could be the graphql artifact's name _or_ a path. In the case where we pass a path this leaves us without a valid variable name to assign the import to. With commonJS we never encounter this because we can put the `require` inline and never assign it to a variable. If we want to swich to ES modules, we need to be able to assing these to variables and that exposes the error. Basically, we end up trying to use the module path as a variable which is not valid.

Reviewed By: rbalicki2

Differential Revision: D42816087

fbshipit-source-id: 1c605477fe93ea543c9765e80b6400f7e0cf7baf