Add new blog post adopted from internal Coinbase blog (#4155)
Summary:
This new post is written in MDX format so I took advantage of a few Docusaurus features to add an image component which makes embedding images a bit better.

I'm not entirely sure why the `Cargo.lock` file needs to be updated, but added it to this PR to try and get things to pass.

Pull Request resolved: https://github.com/facebook/relay/pull/4155

Reviewed By: voideanvalue

Differential Revision: D42551958

Pulled By: davidmccabe

fbshipit-source-id: cfdc3ece76625b5798e0bdf4dbbda758f78e2c62