Log event when we find an operation to suspend on
Summary:
As part of our discussion around exploring a Reactive Key Value store architecture, we identified that OperationTracker would be incompatible in its current form.

As a baseline we would like to be able to collect data showing how effective OperationTacker is today. One way to measure that would be: How often does it identify an operation on which we could suspend that is _not_ the fragment's parent operation. This log event will allow us to track that statistic in prod.

Reviewed By: rbalicki2

Differential Revision: D43294511

fbshipit-source-id: 6e6f648234dc95b2b71ecd01b918a16df47a9c87