fix jest describe.each type (#4236)
Summary:
Pull Request resolved: https://github.com/facebook/relay/pull/4236

This was buggy because it used TArguments as a rest element, even though it was a subtype of mixed — so could be anything. You shouldn’t spread a string/number/etc. Flow 0.201 now complains about this.

The change to use Array<any> is what newer versions of the jest defs already do.

Reviewed By: SamChou19815

Differential Revision: D43810723

fbshipit-source-id: da44563ef7d4317c7895152151e97462ef7384ad