Rerun relay compiler
The project comes with precompiled relay fragments that do not work when the student who is just starting this tutorial tries to fetch the top query without having ant fragments defined. To make it work, we need to delete the precompiled queries and re compile. The student also does not know the command to do that since it is not mentioned anywhere since they may not know to look for it in the `package.json` file. These specifics are important to unblock the student at this stage.