Rerun relay compiler (#4288)
Summary:
The project comes with precompiled relay fragments that do not work when the student who is just starting this tutorial tries to fetch the top query without having any fragments defined. To make it work, we need to delete the precompiled queries and re compile. The student also does not know the command to do that since it is not mentioned anywhere and they may not know to look for it in the `package.json` file. These specifics are important to unblock the student at this stage.

Pull Request resolved: https://github.com/facebook/relay/pull/4288

Reviewed By: voideanvalue

Differential Revision: D46361100

Pulled By: captbaritone

fbshipit-source-id: 3a6775002366ece555e54f605507e88d293ef471