non-nullable __relay_model_instance
Summary: I don't think it make sense for a resolver for a subfield on model type to have to handle the model instance being nullable. If someone wants that, they should make the flow/typescript type for the model be nullable (e.g. `TodoModelType` in `relay-runtime/store/__tests__/resolvers/TodoModel.js`)

Reviewed By: captbaritone

Differential Revision: D46488624

fbshipit-source-id: c697b86f3b3ca403bed3b4b3933cb3448d3e74e6