Extend multiple schemas in parallel
Summary:
Instead of generating/validating schemas 1-by-1, let's generate them simultaneously. If we do it fast enough, we can replace WWW's parallelism with Rust's, which may in theory be faster.

I had to remove parallelism from `schema-validate` as it was making the whole thing ~7 seconds slower. Tried experimenting with limiting the concurrency and got the same results - it works best when validation is not parallelized at all.

 Removing parallelization in schema-validate which turns out to hurt performance.

Reviewed By: gary-z

Differential Revision: D47136978

fbshipit-source-id: e822e9b41dd02b4f00673e6f66aca66d3caa5d32