Fix TypeScript before continuing in tutorial (#4355)
Summary:
While following the tutorial, I was puzzled as the project did not compile at the point where in the text it read: ‘At this point, you should see a story fetched from the server’.

It turned out that the TypeScript error is described below and it is expected. I think it should be placed sooner so it fits the flow.

I decided to move the details about the types to a Deep dive section as it did not fit well elsewhere.

Pull Request resolved: https://github.com/facebook/relay/pull/4355

Reviewed By: voideanvalue

Differential Revision: D47339305

Pulled By: captbaritone

fbshipit-source-id: 901d99c8d783323a51b0e6cddc1a7606bc73f67a