Fix compiler issue with stale artifacts for Resolvers in watch mode (#4415)
Summary:
The Issue:
When running the compiler in watch mode, the compiler may produce
artifacts for relay resolvers.

For example:
```
/** RelayResolver MyType */
```
This code would create fragments for the model resolver, including
an identifier, a model instance fragment, etc.

However, a problem arises in watch mode when we rename the type to:

```
/** RelayResolver  MyNewType */
```

In this case, the old artifacts are not automatically deleted. The issue lies in
the compiler's inability to establish a connection between the resolver
definition code and the artifacts it produces.

To resolve this, we need to establish that connection.

The solution:

The relationship between the source documents (fragments, queries, and now resolvers)
and the generated artifacts is represented in the artifact_map. In this map, the key
corresponds to the document's name (fragment or query), and the value is a
list of the generated artifacts.

For resolvers, we will use the resolver's hash (MD5 of the source) as the key
in the map, and the associated value will be the list of generated artifacts.

The solution involves passing this hash to the artifact generation *place*
and ensuring the proper handling of the removal of outdated artifacts.

Pull Request resolved: https://github.com/facebook/relay/pull/4415

Test Plan:
**Watch mode**

- Add new resolver /** RelayResolver MyType **/
- New files are created for the resolver (id, model_instance)
- Change the name to  /** RelayResolver MyNewType **/
- New files are created, files for MyType are deleted.

 ---

**Non-watch Compilation (no saved state)**

Run 1:
- Add new resolver /** RelayResolver MyType **/
- New files are created for the resolver (id, model_instance)

Rename - Change the name to  /** RelayResolver MyNewType **/

Run 2:
- New files are created, files for MyType are deleted.

 ---

**Non-watch Compilation (with saved-state)**

Pre-run:

- Add new resolver /** RelayResolver MyType **/
- New files are created for the resolver (id, model_instance)
- Commit changes.

Run compiler and create saved state:

```
./scripts/compiler.sh xplat build --export-state=/tmp/compiler-state
```

Run compiler with saved state

```
./scripts/compiler.sh xplat build --import-state=/tmp/compiler-state
```

- No changes, quick compilation.
- Update file: Change the name to  /** RelayResolver MyNewType **/

Run compiler with the saved state:

Old files deleted, new files created:
 {F1076577864}

 ---

**Build with saved-state + dirty resolvers artifact**

See the same steps for for preparation of the saved-state.

Also run:

- Update/Remove generated artifact for the Resolver
- Run compiler with saved-state
- Artifacts are restored to the correct version.

Reviewed By: captbaritone

Differential Revision: D48588439

Pulled By: alunyov

fbshipit-source-id: 72fbebbf79eed0d512d24575601b203c3aee85a1