Update the name of the generated files for resolvers to include project name as prefix (#4418)
Summary:
We have two methods that generate files for resolvers. This diff is passing the `project_name` to these methods, and we're using it as a prefix for fragment/operation name. These fragment/operation names will be used to generate artifacts.

With this prefix we will prevent possible collisions of generated artifacts for projects that have resolvers with the same name.

Pull Request resolved: https://github.com/facebook/relay/pull/4418

Test Plan:
- Updated Snapshot Tests
- Run compiler in OSS
- Run compiler on XPLAT + Adding new model resolver

Also confirmed that there is no flow errors with this change on the Zenon refactoing: D48738810

Reviewed By: captbaritone

Differential Revision: D48608429

Pulled By: alunyov

fbshipit-source-id: bd59c426458e04ce41a8d416a576f497aeb443d3