Update connections-pagination.md (#4423)
Summary:
There are mistakes in Step 2 of the infinite scrolling tutorial.

- `newsfeedStories` is under `viewer`
- We are using `NewsfeedQuery` not `NewsfeedFragment`

In Step 5, the update is shown to a component `NewsfeedContents`, but that doesn't exist in the cloned repo. Is it something we are supposed to create when following along? I don't see this called out.

Step 6 then update the same `NewsfeedContents` to be different again, but in a way that seems as if the wrong component is being used in the example code? It's all a little confusing. This PR fixes a couple of mistakes, but I think the page needs a proper review by someone more familiar with the tutorial.

Pull Request resolved: https://github.com/facebook/relay/pull/4423

Reviewed By: voideanvalue

Differential Revision: D48790897

Pulled By: alunyov

fbshipit-source-id: d94ab1706b34452548cc48e8097b272d9b755141