Refactoring of the type generation/type assertion for provided variables to support internal use case with shared artifacts. (#4429)
Summary:
The main goal is to separate `codegen` and `typegen` for provided variables. Currently, there is an implicit dependency between `typegen` and `codegen` where were generating code for provided variables. The `codegen` expects the local type for provided variables to always be generated as `type ProvidedVariables` so we cam enable inline type assertion as

```
var providedVariables: ProvidedVariables = { ... object  with provided variables ... }
```

However, this approach doesn't work well with the `skip_types` option, where the `typegen` block for an operation is completely skipped (without the `skip_types` we would generate a type for `ProvidedVariables` in it), but the type inline assertion still remains in the `relay-codegen`.

The proposed solution is to make this dependency explicit by passing the generated provided variables object to the method responsible for generating types. Additionally, the type assertion should only be generated in the `relay-typegen`.

 ---
Pull Request resolved: https://github.com/facebook/relay/pull/4429

Test Plan: TBD

Differential Revision: D48801823

Pulled By: alunyov

fbshipit-source-id: 556098bb2ad2457d451b5e77bc38fdb568df2d08