Refactoring of the type generation/type assertion for provided variables to support internal use case with shared artifacts
Summary:
The main goal is to separate `codegen` and `typegen` for provided variables (as much as possible). Currently, there is an implicit dependency between `typegen` and `codegen` where were generating code for provided variables. The `codegen` expects the local type for provided variables to always be generated as `type ProvidedVariables` so we can enable inline type assertion as

```
var providedVariables: ProvidedVariables = { ... object  with provided variables ... }
```

However, this approach doesn't work well with the `skip_types` option, where the `typegen` block for an operation is completely skipped (without the `skip_types` we would generate a type for `ProvidedVariables` in it), but the type inline assertion still remains in the `relay-codegen`. And the type assertion for provided variables is also happening outside of the `export type` blocks for the query.

The proposed solution is to make this dependency explicit by passing the generated provided variables object to the method responsible for generating types. Also in the codegen we would directly generate provided variables in the request parameters, and not rely on the `top_level_statements`.

Reviewed By: captbaritone

Differential Revision: D48801823

fbshipit-source-id: 59d56f1e0453c5866a02b1f1d9bbd57fd3083245